sdks/python/pyproject.toml
[tool.poetry]
name = "julep"
version = "0.4.0-rc0"
description = "Julep is a platform for creating \"persistent\" agents which can do long-horizon tasks"
authors = ["Julep Developers <developers@julep.ai>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
httpx = "^0.27.2"
python-dateutil = "^2.9.0.post0"
attrs = "^24.2.0"
cattrs = "^24.1.0"
beartype = "^0.18.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
poethepoet = "^0.28.0"
pytype = "^2024.4.11"
ward = "^0.68.0b0"

[tool.poe.tasks]
codegen = """
openapi-python-client generate \
  --path ../../openapi.yaml \
  --overwrite \
  --meta=none \
  --config ./config.yaml \
  --output-path julep/sdk
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
